+++ srcs/game/utils_2.c
@@ -3,21 +3,21 @@
 /*                                                        :::      ::::::::   */
 /*   utils_2.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hmoukit <hmoukit@student.42.fr>            +#+  +:+       +#+        */
+/*   By: asayad <asayad@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/24 21:43:32 by hmoukit           #+#    #+#             */
-/*   Updated: 2024/12/28 02:38:09 by hmoukit          ###   ########.fr       */
+/*   Updated: 2025/01/14 11:27:56 by asayad           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <cub3d.h>
 
-double	deg2rad(double angle_deg)
+float	deg2rad(float angle_deg)
 {
 	return (angle_deg * PI / 180);
 }
 
-double	rad2deg(double angle_rad)
+float	rad2deg(float angle_rad)
 {
 	return ((angle_rad / PI) * 180);
 }
